Bandit fix recipient card getting stuck on sending state
The recipient card can get stuck on sending files state if the other
device cancels the transfer mid-way through. This happens due to the
Disconnected message in that instance not containing the value to
differentiate Outbound/Inbound and so the message gets dropped.

The bandit fix is to assume that message is Outbound.