Fix crash on dropping folders onto the main page
Ideally the filtering would be done using the `formats` property of
DropTarget, but mime-type doesn't seem to work, and then there's the
`accept` signal that allows you to accept or reject to be dropped items
based on the returned boolean, but there you're only given the
gdk::Drop type, and it has a method to get the mime-types of the about to
be dropped files, but they were found to be the exact same no matter the
file or folder being  dropped , so can't differentiate between files and
folder from that.

So the only solution for now is to filter just before adding the items
to the selected files model.