Allow whitelisting Packet in fire-walled environments
By default, whatever available network port is used, but this makes it
hard to whitelist the app since it may use any port at random.

To solve this, expose an option in preferences which lets the user set a
static custom network port of the user's choosing, which they could
whitelist in their firewall system.

Fixes #12