Fix crash on badly formatted locale strings
Use `unwrap_or_else` with a message instead of `unwrap` on `formatx!`

Since translated locale strings are user inputs, crashing on badly
formatted strings doesn't come across as a good default.

Ideally we'd use the english locale string as fallback here, but
that'd take some more effort.

See #21 for more on this.