Fix received shared text being cut-off for small text
This is due to the logic put into place for removing the quotes from the
strings that we receive from the rqs_lib. Apparently the quoting is only
done for large strings, not ones that are small enough.

Fix this on the rqs_lib side as well since right now this removing of
the quotes is just a patch work.

Fixes #23