Improve text payload cleaning logic to remove only `""\n`
To not remove `"` unnecessarily. The text may contain `"` within the
quotes and shouldn't be removed as they're part of the original text.

The only problem with the unquoting approach is that sometimes when the
sender sends the text unquoted and its the original text instead that
has the `""\n`, it'll be wrongly removed.

Extension of 370ba2f.