Fix some idle recipient cards becoming non-activatable on send on other
When multiple recipient cards are available to pick from, picking one of
them may wrongfully set some other cards as non-activatable when only
the one that was clicked on was supposed to become non-activatable.

This happens due of my genius of trying to "cache" the ListBoxRow to
reduce the iterations to get the corresponding row for the model item.
But, the ListBoxRow can be freely reassigned to some other model's
widget... This is where all hell breaks loose.

I wasted so much time trying to debug this... :( ughhhghghhgh