Fix panic on selecting custom folder if app has host access to said path
This is an issue in logic, where it was assumed that the path given by
FileChooser will be the mounted path and not the host path itself. And,
so the logic breaks in the case where user had given host access to the
folder they selected.

Assume the path given by FileChooser to be host path (persistent path)
if host path xattr is empty. This is fine even in case where the xattr
is empty due to there being a very old version of desktop portals on the
host system, since the "Downloads folder" will get reverted if the app
doesn't have access to it on the next start.