Fix received files not being available in custom folders like `/tmp`
This is due files ending up being saved in the `/tmp` of the sandbox.
If xattr host path exists and the paths exists on the FS as well, then
it doesn't necessarily mean that the host path will refer to the host
file system, like in here, where it's just a path in the sandbox.

The solution is to simply consume the path provided by FileChooser.

Path provided is host path if the app has been granted host access to it
via `--filesystem`. Otherwise, it's a mounted path.

Now, there's an issue with the vscode-flatpak extension where while
running the app through it, the path given by FileChooser is always a
mounted path. This was the reason for all sorts of wrong logic being
implemented while working on this feature.