Fix unnecessary looking for mDNS advertisements on device rename
An issue with the force param of the start_mdns_discovery method.

When force param is present, we're only supposed to start mDNS discovery
task based on a bool value. So start if true, else not.

So the param type should've been Option<bool> instead of Option<()>

This issue leads to unnecessarily looking for mDNS advertisements when
we aren't even sending files.