Add desktop notifications for incoming transfers
Using desktop portals.
https://docs.rs/ashpd/latest/ashpd/index.html

And, have an auto-decline after 10s since there's no way to prevent the
user from closing the notification, and we can't know on our end whether
the user has closed it or not. So we need to have some sort of timeout,
otherwise the request will forever be pending and user won't be able to
receive any other request, unless they open the app and cancel/accept
the request.