Expose D-Bus API to send list of file paths to Packet
This will be used by the Nautilus extension written in python to send
the file path of files selected by the user in the file manager.

And, Packet can open the received files and show them in the selected
files page with the "Select Recipients" dialog opened for convenience.

---

Multiple approaches were considered for sending a list of paths from
the nautilus extension to the instance of app, or to launch the app
first if it wasn't running, such as:

1. Figuring out the exec command from the Desktop file using
Gio.DesktopAppInfo and running the command with some `--send-files`
arg that takes in a list of paths. But, the issue being in passing
these options from a remote instance to the primary instance of the
app while using the same binary instead. `is_remote()` couldn't be
used to set the HANDLES_COMMAND_LINE application flag dynamically.

2. Having our own D-Bus API using zbus, but that'd require us to use a
separate service name and file than the one GApplication will be
using.

3. Or, to export our Dbus objects using DBusConnection, unfortunately
the documentation is quite lackluster. But, thankfully it allows for
exporting ActionGroup which is easy enough to do even though the
resulting generated API is not too ergonomic.

\#3 is currently how it's implemented. The ActionGroup is exported
under `/<APP_ID Path>/Share` and the action can be called from the
`Activate` method over the interface `org.gtk.Actions` with parameters
such as:

```
('send-files', [<['~/file_1', '~/file_2']>], {})
```