Add Nautilus python extension for "Send with Packet" menu item
This is done using the ActionGroup D-Bus API exposed by Packet, which
was added back in the previous commit.

A list of file paths as strings is sent to the `send-files` Action.

This python script will need to be copied or symlinked to the
appropriate directory in order to be accessible by Nautilus.

An interface to do that is currently missing from the application.